Update database configuration and connection strings
Refactor `YuzuIdentityDbContextFactory.cs` to improve the loading of configuration files and set the base path to the current directory. Update `DefaultConnection` in `appsettings.Development.json` and `appsettings.json` to point to a new cloud-hosted PostgreSQL database. Clear the `ConnectionString` in `DataStorageConfig` and restructure the JSON files to include necessary logging configurations.